完善商品添加和人员添加功能

+++ cart-service/src/main/java/com/hmall/cart/service/impl/CartServiceImpl.java
@@ -140,7 +140,7 @@ public class CartServiceImpl extends ServiceImpl<CartMapper, Cart> implements IC
     }
 
     private void checkCartsFull(Long userId) {
-        int count = lambdaQuery().eq(Cart::getUserId, userId).count();
+        Long count = lambdaQuery().eq(Cart::getUserId, userId).count();
         if (count >= cartProperties.getMaxItems()) {
             throw new BizIllegalException(
                     StrUtil.format("用户购物车课程不能超过{}", cartProperties.getMaxItems()));
@@ -148,7 +148,7 @@ public class CartServiceImpl extends ServiceImpl<CartMapper, Cart> implements IC
     }
 
     private boolean checkItemExists(Long itemId, Long userId) {
-        int count = lambdaQuery()
+        Long count = lambdaQuery()
                 .eq(Cart::getUserId, userId)
                 .eq(Cart::getItemId, itemId)
                 .count();
